add new mappings

+++ src/main/java/edu/daniel/lordoftheringsbd/Services/IPosesionService.java
@@ -2,7 +2,6 @@ package edu.daniel.lordoftheringsbd.Services;
 
 import java.util.List;
 
-import edu.daniel.lordoftheringsbd.entities.Personaje;
 import edu.daniel.lordoftheringsbd.entities.Posesion;
 
 public interface IPosesionService {
@@ -10,7 +9,7 @@ public interface IPosesionService {
 
     public Posesion findById(Integer id);
 
-    public void borrarPosesion(Integer id);
+    public void borrarPosesion(Long id);
 
     public Posesion guardarPosesion(Posesion pos);
